check for file existence before deleting tmp file

+++ src/java/org/apache/cassandra/service/opstate/KeyspaceTableOpStatePersister.java
@@ -149,7 +149,9 @@ public class KeyspaceTableOpStatePersister
         }
         finally
         {
-            Files.delete(tmpFile.toPath());
+            if (tmpFile.exists()) {
+                Files.delete(tmpFile.toPath());
+            }
         }
     }
